refactor: split derived maintenance and session-product refresh roots
Replace mixed async derived and session-product refresh roots with explicit action, product, maintenance, stats, update, and delete control bands.
