+++ polylogue/storage/backends/async_sqlite_derived_products.py
@@ -1,9 +1,7 @@
-"""Derived-model write/query methods for the async SQLite backend."""
+"""Derived session-product writes for the async SQLite backend."""
 
 from __future__ import annotations
 
-from polylogue.storage.backends.queries import action_events as action_events_q
-from polylogue.storage.backends.queries import maintenance_runs as maintenance_runs_q
 from polylogue.storage.backends.queries import (
     session_product_profile_writes as session_product_profiles_q,
 )
@@ -13,11 +11,7 @@ from polylogue.storage.backends.queries import (
 from polylogue.storage.backends.queries import (
     session_product_timeline_writes as session_product_timelines_q,
 )
-from polylogue.storage.backends.queries import stats as stats_q
 from polylogue.storage.store import (
-    ActionEventRecord,
-    MaintenanceRunRecord,
-    MessageRecord,
     SessionPhaseRecord,
     SessionProfileRecord,
     SessionWorkEventRecord,
@@ -25,33 +19,8 @@ from polylogue.storage.store import (
 )
 
 
-class SQLiteDerivedMixin:
-    """Derived action/product/maintenance methods for ``SQLiteBackend``."""
-
-    async def replace_action_events(
-        self,
-        conversation_id: str,
-        records: list[ActionEventRecord],
-    ) -> None:
-        """Replace durable action-event rows for one conversation."""
-        async with self._get_connection() as conn:
-            await action_events_q.replace_action_events(
-                conn,
-                conversation_id,
-                records,
-                self._transaction_depth,
-            )
-
-    async def get_action_events(self, conversation_id: str) -> list[ActionEventRecord]:
-        """Get durable action-event rows for one conversation."""
-        return await self.queries.get_action_events(conversation_id)
-
-    async def get_action_events_batch(
-        self,
-        conversation_ids: list[str],
-    ) -> dict[str, list[ActionEventRecord]]:
-        """Get durable action-event rows for multiple conversations."""
-        return await self.queries.get_action_events_batch(conversation_ids)
+class SQLiteDerivedProductsMixin:
+    """Derived durable session-product methods for ``SQLiteBackend``."""
 
     async def replace_session_profile(
         self,
@@ -107,29 +76,5 @@ class SQLiteDerivedMixin:
                 self._transaction_depth,
             )
 
-    async def record_maintenance_run(
-        self,
-        record: MaintenanceRunRecord,
-    ) -> None:
-        """Persist one maintenance lineage record."""
-        async with self._get_connection() as conn:
-            await maintenance_runs_q.record_maintenance_run(
-                conn,
-                record,
-                self._transaction_depth,
-            )
 
-    async def upsert_conversation_stats(
-        self,
-        conversation_id: str,
-        provider_name: str,
-        messages: list[MessageRecord],
-    ) -> None:
-        """Upsert precomputed per-conversation aggregate stats."""
-        async with self._get_connection() as conn:
-            await stats_q.upsert_conversation_stats(
-                conn, conversation_id, provider_name, messages, self._transaction_depth
-            )
-
-
-__all__ = ["SQLiteDerivedMixin"]
+__all__ = ["SQLiteDerivedProductsMixin"]
